A safer wrapper script for launching the beacon node containers in docker

+++ docker/hosts.nims
@@ -1,4 +1,5 @@
-import strformat
+import
+  strformat
 
 var
   serverCount = 10
@@ -11,8 +12,7 @@ var
   validatorsPerServer = systemValidators div serverCount
   validatorsPerNode = validatorsPerServer div instancesCount
 
-  network = "testnet1"
-  buildFlavour = "libp2p_spec"
+  network = "testnet0"
 
 type
   ContainerCmd = enum
@@ -22,23 +22,23 @@ type
 iterator containers: tuple[cmd: ContainerCmd, node, container: string, firstValidator, lastValidator: int] =
   for i in 0 ..< serverCount:
     let baseIdx = userValidators + i * validatorsPerServer
-    for j in 0 ..< instancesCount:
+    for j in 1 .. instancesCount:
       let firstIdx = baseIdx + j * validatorsPerNode
       let lastIdx = firstIdx + validatorsPerNode - 1
-      yield (restart, &"nimbus-node-0{i}", &"beacon-node-{buildFlavour}-{network}-{j}", firstIdx, lastIdx)
-    for j in instancesCount .. 2:
-      yield (stop, &"nimbus-node-0{i}", &"beacon-node-{buildFlavour}-{network}-{j}", -1, -1)
+      yield (restart, &"nimbus-node-0{i}", &"beacon-node-{network}-{j}", firstIdx, lastIdx)
 
-for i in 0 ..< serverCount:
-  let baseIdx = userValidators + i * validatorsPerServer
-  for j in 1 .. instancesCount:
-    let firstIdx = baseIdx + (j - 1) * validatorsPerNode
-    let lastIdx = firstIdx + validatorsPerNode - 1
-    let dockerPath = &"/docker/beacon-node-{buildFlavour}-{network}-{j}/data/BeaconNode/{network}"
-    echo &"ssh nimbus-node-0{i} 'sudo mkdir -p {dockerPath}/validators && sudo rm -f {dockerPath}/validators/* && " &
-                               &"sudo ~/nimbus/vendor/nim-beacon-chain/scripts/download_validator_keys.sh {network} {firstIdx} {lastIdx} {dockerPath} && " &
-                               &"sudo chown dockremap:docker -R {dockerPath}'"
+when false:
+  for i in 0 ..< serverCount:
+    let baseIdx = userValidators + i * validatorsPerServer
+    for j in 1 .. instancesCount:
+      let firstIdx = baseIdx + (j - 1) * validatorsPerNode
+      let lastIdx = firstIdx + validatorsPerNode - 1
+      let dockerPath = &"/docker/beacon-node-{network}-{j}/data/BeaconNode/{network}"
+      
+      echo &"ssh nimbus-node-0{i} 'sudo mkdir -p {dockerPath}/validators && sudo rm -f {dockerPath}/validators/* && " &
+                                 &"sudo ~/nimbus/vendor/nim-beacon-chain/scripts/download_validator_keys.sh {network} {firstIdx} {lastIdx} {dockerPath} && " &
+                                 &"sudo chown dockremap:docker -R {dockerPath}'"
 
-# for c in containers():
-#  echo &"ssh {c.node} docker {c.cmd} {c.container}"
+for c in containers():
+  echo &"ssh {c.node} docker {c.cmd} {c.container}"
